+++ tests/integration/integration.test.ts
@@ -16,6 +16,7 @@ it('outputs accessibility violation information for the demo app', (done) => {
     expect(normalizedStdout).toMatchSnapshot();
     done();
   });
+  // @ts-expect-error Getting Mocha instead of Jest types...
 }, 120000);
 
 it('filters the components to run', (done) => {
@@ -27,6 +28,7 @@ it('filters the components to run', (done) => {
     expect(normalizedStdout).toMatchSnapshot();
     done();
   });
+  // @ts-expect-error Getting Mocha instead of Jest types...
 }, 120000);
 
 it('fails only specific impact levels if specified', (done) => {
@@ -38,6 +40,7 @@ it('fails only specific impact levels if specified', (done) => {
     expect(normalizedStdout).toMatchSnapshot();
     done();
   });
+  // @ts-expect-error Getting Mocha instead of Jest types...
 }, 120000);
 
 /**
@@ -49,8 +52,8 @@ function normalize(input: string) {
   const specTimePattern = /\s*\([\d.]+m?s\)/g;
   /** File system paths. For example, `/path/to/some/file */
   const cwdPattern = new RegExp(process.cwd(), 'g');
-  /** Line numbers from stack trace paths. For example, `.js:20:55` */
-  const lineNumbersPattern = /\.js:\d+:\d+/g;
+  /** Line numbers from stack trace paths. For example, `timers:20:55` */
+  const lineNumbersPattern = /:\d+:\d+/g;
   /** storybook-start outputs an info dialog at the beginning this removes those lines */
   const storybookStartedFrame = /(│.*│\n|[─╭╮╰╯]+?)/g;
   /** webpack outputs a hash and build time */
@@ -62,7 +65,7 @@ function normalize(input: string) {
   return input
     .replace(specTimePattern, '')
     .replace(cwdPattern, '')
-    .replace(lineNumbersPattern, '.js')
+    .replace(lineNumbersPattern, '')
     .replace(storybookStartedFrame, '')
     .replace(webpackBuilt, 'webpack built')
     .replace(webpackBuildTimeVariations, '');