+++ tests/unit/Result.test.ts
@@ -1,4 +1,5 @@
-import { isPassing, Result } from '../../src/Result';
+import dedent from 'ts-dedent';
+import { isPassing, formatFailureResult, type Result } from '../../src/Result';
 
 describe('isPassing', () => {
   it('returns true when there are no violations', () => {
@@ -95,5 +96,45 @@ describe('isPassing', () => {
     };
 
     expect(isPassing(result, ['critical'])).toEqual(true);
+  });
+});
+
+describe('formatFailureResult', () => {
+  it('pretty prints the failures', () => {
+    const result: Result = {
+      violations: [
+        {
+          description: 'Ensures buttons have discernible text',
+          help: 'Buttons must have discernible text',
+          helpUrl: 'https://dequeuniversity.com/rules/axe/4.0/button-name',
+          id: 'button-name',
+          tags: [],
+          nodes: [
+            {
+              html: '<button></button>',
+              target: [],
+              any: [],
+              all: [],
+              none: [],
+              failureSummary: "Nope, that is not right, because\nIt's not",
+            },
+          ],
+        },
+      ],
+    };
+
+    expect(formatFailureResult(result)).toEqual(dedent`
+      Detected the following accessibility violations!
+
+      1. button-name (Buttons must have discernible text)
+
+         For more info, visit https://dequeuniversity.com/rules/axe/4.0/button-name.
+
+         Check these nodes:
+
+         - html: <button></button>
+           summary: Nope, that is not right, because
+                    It's not
+    `);
   });
 });