+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 5.0.1 (2022-05-20)
 
-- [maintenance] Upgrade all dependencies
+- [breaking] Remove `--format` option
 - [fix] Access static storybook via file server so everything works with storyStoreV7 (fixing #51)
+- [maintenance] Upgrade all dependencies
+- [maintenance] Use MochaJS internally for running the "tests"
 
 ## 5.0.0 (2021-11-29)